+++ pages/pages/dataGridPage.tsx
@@ -1,7 +1,30 @@
 import Box from '../../src/box';
 import DataGrid from '../../src/components/dataGrid';
-import { GridDef } from '../../src/components/dataGrid/dataGridContract';
+import { GridDefinition } from '../../src/components/dataGrid/dataGridContract';
 import Grid from '../../src/components/grid';
+import Data from '../data/MOCK_DATA.json';
+
+interface DataType {
+  first_name: string;
+  last_name: string;
+  age: number;
+  email: string;
+  country: string;
+  favorite_color: string;
+  gender: string;
+  job_title: string;
+  ssn: string;
+  birthdate: string;
+  phone_number: string;
+  street_address: string;
+  city: string;
+  username: string;
+  credit_card_number: string;
+  salary: number;
+  company_name: string;
+  language: string;
+  currency_code: string;
+}
 
 interface Person {
   firstName: string;
@@ -162,34 +185,43 @@ const data: Person[] = [
 ];
 
 export default function DataGridPage() {
-  var def: GridDef<Person> = {
-    columns: [
-      { key: 'visits' },
-      { key: 'status' },
-      {
-        key: 'employee',
-        columns: [
-          { key: 'age', columns: [{ key: 'month' }, { key: 'year' }] },
-          { key: 'name', columns: [{ key: 'fullName', columns: [{ key: 'firstName' }, { key: 'lastName' }] }] },
-        ],
-      },
-      { key: 'progress' },
-    ],
-  };
-
-  var def2: GridDef<Person> = {
-    columns: [
-      {
-        key: 'employee',
-        columns: [{ key: 'firstName' }, { key: 'lastName' }],
-      },
-      { key: 'visits' },
-      { key: 'status' },
-
-      { key: 'progress' },
-    ],
+  var def: GridDefinition<DataType> = {
     pagination: true,
+    columns: [
+      { key: 'name', columns: [{ key: 'first_name' }, { key: 'last_name' }] },
+      { key: 'age' },
+      { key: 'email' },
+      { key: 'country' },
+      { key: 'favorite_color' },
+      { key: 'gender' },
+      { key: 'job_title' },
+      { key: 'ssn' },
+      { key: 'birthdate' },
+      { key: 'phone_number' },
+      { key: 'street_address' },
+      { key: 'city' },
+      { key: 'username' },
+      { key: 'credit_card_number' },
+      { key: 'salary' },
+      { key: 'company_name' },
+      { key: 'language' },
+      { key: 'currency_code' },
+    ],
   };
+
+  // var def2: GridDef<Person> = {
+  //   columns: [
+  //     {
+  //       key: 'employee',
+  //       columns: [{ key: 'firstName' }, { key: 'lastName' }],
+  //     },
+  //     { key: 'visits' },
+  //     { key: 'status' },
+
+  //     { key: 'progress' },
+  //   ],
+  //   pagination: true,
+  // };
 
   return (
     <Box>
@@ -197,8 +229,8 @@ export default function DataGridPage() {
         DataGrid
       </Box>
 
-      <DataGrid data={data} def={def} />
-      <DataGrid data={data} def={def2} mt={10} />
+      <DataGrid data={Data} def={def} />
+      {/* <DataGrid data={data} def={def2} mt={10} /> */}
     </Box>
   );
