+++ src/components/dataGrid.tsx
@@ -16,27 +16,38 @@
 // datagrid container
 
 import Box, { BoxProps } from '../box';
-import { GridDef } from './dataGrid/dataGridContract';
+import { GridDefinition } from './dataGrid/dataGridContract';
 import Grid from './grid';
-import useGridData from './dataGrid/useGridData';
+import useGridData2 from './dataGrid/useGridData2';
 import Flex from './flex';
-import DataGridCell from './dataGrid/dataGridCell';
+import DataGridCell2 from './dataGrid/dataGridCell2';
 import DataGridPagination from './dataGrid/dataGridPagination';
+import { useMemo } from 'react';
 
 interface Props<TRow> extends BoxProps {
   data: TRow[];
-  def: GridDef<TRow>;
+  def: GridDefinition<TRow>;
 }
 
 export default function DataGrid<TRow extends {}>(props: Props<TRow>) {
   const { def } = props;
-  const grid = useGridData(props);
+  const grid = useGridData2(props);
 
   const sizes = grid.headerColumns.reduce<Record<string, string>>((acc, c) => {
     acc[`--${c.key}-${c.pinned ?? ''}-width`] = typeof c.inlineWidth === 'number' ? `${c.inlineWidth}px` : (c.inlineWidth ?? 'unset');
 
     return acc;
   }, {});
+
+  const renderRows = useMemo(() => {
+    return grid.rows.map((row) => (
+      <Box key={row.key} display="contents" className="grid-row">
+        {row.cells.map((cell) => (
+          <DataGridCell2 key={`${cell.key}${cell.pinned}`} row={row} cell={cell} grid={grid} />
+        ))}
+      </Box>
+    ));
+  }, [grid.rows]);
 
   return (
     <Box component="dataGrid" b={1} borderRadius={1} overflow="hidden" {...props} style={sizes}>
@@ -45,13 +56,7 @@ export default function DataGrid<TRow extends {}>(props: Props<TRow>) {
       </Box>
       <Box overflow="auto" height={112} bb={1}>
         <Grid style={{ gridTemplateColumns: grid.gridTemplateColumns }} userSelect={grid.isResizeMode ? 'none' : undefined}>
-          {grid.rows.map((row) => (
-            <Box key={row.key} display="contents" className="grid-row">
-              {row.cells.map((cell) => (
-                <DataGridCell key={`${cell.key}${cell.pinned}`} row={row} cell={cell} grid={grid} />
-              ))}
-            </Box>
-          ))}
+          {renderRows}
         </Grid>
       </Box>
       <Flex p={3} jc="space-between" height={10}>
